Moved the Form tests to a dedicated namespace in the testsuite
